Greg: I switched the product rule to (fg)'=f'g+fg', and a corresponding change for the quotient rule.
The rule used to be: (fg)'=g'f+f'g. This is how its in the book - and that's the very very morally wrong way of arranging the multiplication. The reason: differentiation is also defined for non-commutative rings (matrices! Lie theory! a large portion of modern mathematics!), and the product rule is exactly as I wrote it (you do not have the right to switch the order of multiplication!).  Quite frankly, the book's messing up the order of multiplicands is burning my Lie theorists' eyes.