Defining new commands \Arcsin, \Arccos, etc. \Arcsin is defined as \newcommand{\Arcsin}{\sin^{-1}} in the standard .sty file. In my personal lectures, \Arcsin is redefined as \renewcommand{\Arcsin}{\arcsin}.
Motivation for the change: 1. This does not in any way change Greg's slides. 2. I strongly believe the notation \sin^{-1} x is very very bad in view of its obvious conflict with the notation \sin^2 x. That is why I refuse to use the \sin^{-1} x notation and instead use \arcsin. 3. Using \Arcsin, both notations can be used by a single one-line renewal of the \Arcsin command. 4. Note that we are not renewing the standard \arcsin command: renewing already standard commands is *very* bad practice (imagine someone pastes your latex code in a larger project which already uses the \arcsin command).